Half-baked solution for ocpp+origin, breakin in deserialization

+++ tasks/chargepoint.py
@@ -1,8 +1,8 @@
 import asyncio
+import calendar
 import datetime
 
 import energyweb
-from energyweb import ExternalData
 
 from tasks.database.dao import DAOFactory
 from tasks.database.elasticdao import ElasticSearchDAOFactory, ElasticSearchDAO
@@ -11,11 +11,28 @@ from tasks.ocpp16.protocol import ChargingStation
 from tasks.ocpp16.server import Ocpp16Server
 
 
-class EVchargerEnergyMeter(energyweb.IntegrationPoint):
-    def read_state(self, *args, **kwargs) -> ExternalData:
-        pass
+class EVchargerEnergyMeter(energyweb.EnergyDevice):
 
-    def write_state(self, *args, **kwargs) -> ExternalData:
+    def __init__(self, service_urls: tuple, manufacturer, model, serial_number, energy_unit, is_accumulated,
+                 latitude=None, longitude=None):
+        self.service_urls = service_urls
+        super().__init__(manufacturer, model, serial_number, energy_unit, is_accumulated, latitude, longitude)
+
+    def read_state(self, *args, **kwargs) -> energyweb.EnergyData:
+        els_dao: ElasticSearchDAO = ElasticSearchDAOFactory('elocity', *self.service_urls).get_instance(ChargingStation)
+        results = els_dao.retrieve_all()
+        results_a = els_dao.query({"bool": {
+            "must_not": {"exists": {"field": 'co2_saved'}},
+            "must": [{"exists": {"field": 'meter_start'}}, {"exists": {"field": 'meter_stop'}}]
+        }})
+        results_b = els_dao.query({"bool": {
+            "must": [{"exists": {"field": 'meter_start'}}, {"exists": {"field": 'meter_stop'}}]
+        }})
+        now = datetime.datetime.now().astimezone()
+        return energyweb.EnergyData(device=self, access_epoch=calendar.timegm(now.timetuple()), raw='',
+                                      energy='', measurement_epoch='')
+
+    def write_state(self, *args, **kwargs) -> energyweb.EnergyData:
         pass
 
 
@@ -27,7 +44,8 @@ class Ocpp16ServerTask(energyweb.Task, energyweb.Logger):
         self._factory = factory
         self.server_address = (host, port)
         self._future = None
-        energyweb.Task.__init__(self, queue, polling_interval=datetime.timedelta(minutes=5), eager=True, run_forever=True)
+        energyweb.Task.__init__(self, queue, polling_interval=datetime.timedelta(minutes=5), eager=True,
+                                run_forever=True)
         energyweb.Logger.__init__(self, 'Ocpp16Server')
 
     class Ocpp16ServerLogger(Ocpp16Server):
@@ -66,9 +84,9 @@ class Ocpp16ServerTask(energyweb.Task, energyweb.Logger):
 
 class ElasticSync(energyweb.Task):
 
-    def __init__(self, queue: dict, service_urls: tuple):
+    def __init__(self, queue: dict, interval: datetime.timedelta, service_urls: tuple):
         self.service_urls = service_urls
-        super().__init__(queue=queue, polling_interval=datetime.timedelta(minutes=3), eager=False, run_forever=True)
+        super().__init__(queue=queue, polling_interval=interval, eager=False, run_forever=True)
 
     async def _prepare(self):
         pass
@@ -79,15 +97,18 @@ class ElasticSync(energyweb.Task):
             try:
                 idd_css = [cs for cs in mem_dao.retrieve_all() if cs.serial_number]
                 idd_css.sort(key=lambda cs: cs.last_seen, reverse=True)
+                if len(idd_css) < 1:
+                    return
                 oldest: ChargingStation = idd_css[0]
-                for cs in idd_css[1:]:
-                    oldest.host = cs.host
-                    oldest.port = cs.port
-                    oldest.tags.update(cs.tags)
-                    oldest.transactions.update(cs.transactions)
-                    oldest.connectors.update(cs.connectors)
-                    oldest.last_heartbeat = cs.last_heartbeat
-                    mem_dao.delete(cs)
+                if len(idd_css) > 1:
+                    for cs in idd_css[1:]:
+                        oldest.host = cs.host
+                        oldest.port = cs.port
+                        oldest.tags.update(cs.tags)
+                        oldest.transactions.update(cs.transactions)
+                        oldest.connectors.update(cs.connectors)
+                        oldest.last_heartbeat = cs.last_heartbeat
+                        mem_dao.delete(cs)
                 mem_dao.update(oldest)
                 oldest.reg_id = oldest.serial_number
                 els_dao.update(oldest)
